Support for reading MATLAB structs and struct arrays
This forced me to remove the previous default behavior for read, where
read(obj, name) defaulted to d_read when obj is a group, and a_read
when obj is an attribute. This change was required because structs are
implemented as groups of datasets, yet we want to be able to read them
as a coherent item.

This change should also prepare us for reading CompositeKinds, for
which representation as a group also makes sense.