Crude support for H5T_COMPOUND datatypes
I was going to wait until Julia has immutables, but it turns out that
reading .mat files with complex numbers requires this
functionality. In MAT_HDF5.jl, the generic "read" function (which is
the part that merits the designation "crude") is bypassed by the
complex support , but the low-level wrappers are used.