Throw error when saving unsupported type parameters
Only allow bitstypes that show as AST literals and make sure that they read
back exactly as they are saved. Use show (instead of print) to preserve as many
Julian type distinctions as we can e.g., 1 vs 0x01. A different implementation
will be required to support custom immutables or things as simple as Int16(1).