Temporarily disable type parameters that are themselves tuples
On 0.4, we can't differentiate between these and tuples of types that
were saved by earlier versions of Julia. In earlier versions of Julia,
such tuples were not bits typed so they would not have been usable
anyway. We can revisit this when we're ready to change the file format
(perhaps as part of #204/#218)