Use jl_set_nth_field for non-pointerfree immutables
Because of the write barrier, what we were doing before wasn't quite
right. Non-pointerfree immutables are heap allocated, but we were
adding references to objects that existed only on the stack. This
does this the "right way," which is in principle a bit slower, but the
cost probably doesn't matter much since reading references is so
expensive.