Reading HDF5Compound now returns properly typed data (#365)
* Can now read in compound datasets with fixed length strings by using an IObuffer

* Compound ready now works with FixedArray types.

* Updated unit test for Compound dsets

* No need to expose offsets, types, since we are directly returning typed data

* Remove extraneous type and offset information for Compound dsets. Use Vectors instead of Arrays.

* Update unit test for 685b14e

* Use Any[] rather than Vector{Any}() as a constructor.

* Explicit type for row data

* Return Vector{HDF5Compound} when reading a vector of compound types.

* Make HDF5Compound immutable.

Including using tuples for the internal element storage. Allows for more
efficient storage of arrays of HDF5Compound.

* Use existing unpad() method to truncate Strings in compound types.

* Be more careful about what we can extract from a compound type.

If an element is not an HDF5BitsKind, then keep it around as raw bytes to let
the user decide how to handle it. We also keep the types around to facilitate
these later manipulations.

* Try to fix tests on julia 0.4

* Immediately grab number of members of HDF5Compound types in hdf5_to_julia_eltype().

I am contemplating also pushing the member names into the type parameters,
because as is, if the user loads a nested compound type, they will not have
sufficient information to reconstruct the "inner" compound type.