Fix heuristic chunking of zero-dimensional arrays
Previously, when calling `heuristic_chunk` with a zero-dimensional array
`x` (i.e. a scalar wrapped by an array, for example `x = ones()`),
heuristic chunking failed with an error, since `shape = size(x)` was an
empty Tuple, and thus `[shape...]` was an `Array{Any,1}` instead of the
expected `Array{Int,1}`.

This is fixed here by creating a variant of `heuristic_chunk` that gets
called when `size(x)` is an empty tuple, i.e., when `x` is a
zero-dimensional array.

In practice, this disables chunking of zero-dimensional arrays (chunking
scalars doesn't make much sense anyway!).