Add test for zero-dimensional arrays
The test writes and reads back a compressed zero-dimensional array (i.e.
a scalar wrapped into an array). Since chunking is not specified in the
tests, writing the compressed array uses `heuristic_chunk`.

Note that the zero-dimensional array is written to HDF5 as a scalar
dataset, since HDF5 doesn't have the notion of zero-dimensional array
(unlike Julia). Therefore, when reading back the data, HDF5 actually
returns a scalar and not a zero-dimensional array.