Fix heuristic chunking of zero-dimensional arrays (#445)
* Fix heuristic chunking of zero-dimensional arrays

Previously, when calling `heuristic_chunk` with a zero-dimensional array
`x` (i.e. a scalar wrapped by an array, for example `x = ones()`),
heuristic chunking failed with an error, since `shape = size(x)` was an
empty Tuple, and thus `[shape...]` was an `Array{Any,1}` instead of the
expected `Array{Int,1}`.

This is fixed here by creating a variant of `heuristic_chunk` that gets
called when `size(x)` is an empty tuple, i.e., when `x` is a
zero-dimensional array.

In practice, this disables chunking of zero-dimensional arrays (chunking
scalars doesn't make much sense anyway!).

* Add test for zero-dimensional arrays

The test writes and reads back a compressed zero-dimensional array (i.e.
a scalar wrapped into an array). Since chunking is not specified in the
tests, writing the compressed array uses `heuristic_chunk`.

Note that the zero-dimensional array is written to HDF5 as a scalar
dataset, since HDF5 doesn't have the notion of zero-dimensional array
(unlike Julia). Therefore, when reading back the data, HDF5 actually
returns a scalar and not a zero-dimensional array.