Simplify data-driven (a|d)_(create|write) implementations.
These are simply high-level wrappers that automatically configure the
datatype and dataspace given some concrete data. As such, we can get
rid of all of the type restrictions on `data` since we can rely on
the later dispatches to `datatype`, `dataspace` (and less so, the
later `h5d_write` specializations) to appropriately restrict the valid
types.