Fix #705, distinguishing 0-size and null arrays (#706)
* Fix #705, distinguishing 0-size and null arrays

HDF5 has the concept of NULL arrays which are distinct from 0-size
arrays. As shown in #705, writing any 0-size array was being translated
to the NULL object during writing, but having 0-length axes is also
perfectly valid. This PR preserves 0-sizes when writing arrays.

To represent the HDF5 NULL object, there was already the
`HDF5.EmptyArray` type defined, but it appears to have been used only
for internal dispatch on the generic read (which then returns a 0-length
vector for NULL objects).

This PR reuses `HDF5.EmptyArray` and elevates it to a full subtype of
`AbstractArray`, which then becomes the returned object for any NULL
datasets/attributes which are read in. This array type carries an
element type but has no contents. This is distinct from Julia's
`Array{T,0}` 0-dimensional arrays because the latter _does_ have a
single element. (I tried just doing duck typing instead of actually
making it a subtype of `AbstractArray`, but you end up playing
whack-a-mole with enough new method definitions that you might as well
just implement the `AbstractArray` interface and take advantage of its
fallbacks once the few methods are defined.)

(One oddity is that Julia's 0-dimensional `Arrays` are turned into HDF5
scalars and will not be read back in as a 0-dimensional array, but (a)
that is the existing behavior, and (b) I haven't thought of any other
way to handle the case.)

* Fix `IndexStyle` signature based on review

Co-authored-by: Mustafa M <mus-m@outlook.com>

* Avoid invalid concatenation of EmptyArrays

* Add short description to docs

Co-authored-by: Mustafa M <mus-m@outlook.com>