Remove default-parameter overloads of low-level `h5?_(open|create)` API
There is one curiosity of of particular note:
For `h5a_create`, the HDF5 docs say that the `acpl_id` argument is
unused, but non-default values are observable even beyond a close and
reopen operation (retrieved with `get_create_properties()`).
Specifically, we supply a non-default property list
`UTF8_LINK_PROPERTIES[]` which has the character encoding changed
explicitly to UTF8, so we maintain (and recommend within the deprecation
binding) the existing behavior despite the docs' recommendation of
passing `H5P_DEFAULT`.