Add support for dimension scales API bindings
This adds bindings to the HDF5 dimension scales API functions (H5DS)

All of the H5DS functions are added, except for `H5DS_ITERATE_SCALES` which I could not figure out how to bind to Julia, and it seemed of dubious use anyway.

`bind_generator.jl` was changed to allow for the H5DS functions to be correctly linked to the high level library instead of the default library.

`gen_wrappers.jl` is changed to add references to the API on the HDF Group's website.