Partially merge opaque reading into generic read framework
Getting and passing along the tag and data together as a combined
`HDF5.Opaque` object makes this not fully mergeable with the generic
read implementation, but we can at least structure the reads similarly.

Also adds a few tests which didn't previously exist (except one minor
use in the remote read test files).