Limit HDF5_jll compatible versions
HDF5_jll follows the versioning scheme of libhdf5 which uses the
second set of digits as their "major" versions. Set the compatibility
to agree with this use of versioning so that HDF5.jl doesn't try to
pull in a newer library (which may be incompatible) without being
explicitly allowed to in the future.