Split out src/HDF5.jl into multiple files
The file was getting huge (>2000 lines), and was difficult to find anything. I've tried to organise it along the lines of how the libhdf5 library itself is arranged. No functionality has changed, though I did need to split out a few Union methods into individual ones depending on where the type was defined.