Add `filter` keyword argument to create_dataset
This allows `create_dataset` to create a dataset with a user-defined
filter by setting the `filter` keyword to a Tuple of integers that
describe and configure the filter.  For example, to use a filter
specified by `filter_id=32013`, `flags=0`, `cd_nelmts=3`, and
`cd_values=(2,0,8)`:

    ds, dt = create_dataset(h5, "data", data,
                            chunk=(256,1,256),
                            filter=(32013,0,2,0,8))

Note that `cd_nelmts` is not included explicitly.  It is inferred from
the length of the Tuple.