Add `filter` keyword argument to create_dataset (#840)
* Add `filter` keyword argument to create_dataset

This allows `create_dataset` to create a dataset with a user-defined
filter by setting the `filter` keyword to a Tuple of integers that
describe and configure the filter.  For example, to use a filter
specified by `filter_id=32013`, `flags=0`, `cd_nelmts=3`, and
`cd_values=(2,0,8)`:

    ds, dt = create_dataset(h5, "data", data,
                            chunk=(256,1,256),
                            filter=(32013,0,2,0,8))

Note that `cd_nelmts` is not included explicitly.  It is inferred from
the length of the Tuple.

* Add H5Z_FLAG_MANDATORY constant

This is useful for user-defined filters.

* Support multiple user-defined filters per dataset

The `filter` keyword argument now accepts `Tuple{Tuple{<:Integer}}` to
enable multiple user-defined filters when creating a dataset.  Passing
`Tuple{<:Integer}` for a single user-defined filter is still supported.

* Add tests for filter keyword arg of create_dataset

* Change some type/names based on code review

Better names are better! :)

* Change how `set_filter` passes Tuple as Ptr{Cuint}

This new way actually allocates fewer bytes and is compatible with
Julia 1.3.

* Add comment about passing cd_values in set_filter

(cherry picked from commit 834141a027ba1ca1d6c2b608a56ce5ba57389e24)