Support BITSHUFFLE option of Blosc >= v0.7.3 (#899)
Blosc 0.7.3 now treats the blosc "shuffle" option as an Integer rather
than a Bool.  This allows the use of all three of blosc's shuffling
options: no shuffling (0), byte shuffling (1), or bit shuffling (2).
The `shuffle` field of the BloscFilter struct is not restricted to Bool,
but it was effectively being converted to Bool via `!=(0)` in the
`blosc_filter()` function of the H5Zblosc plugin before passing to the
`blosc_compress()` function.  This prevented the use of bit shuffling
because `2 != 0` is `true` which is the same as 1 (byte shuffling).

The H5Zblosc plugin no longer converts the given value to Bool and also
validates the value against the three supported shuffle option values.
These values are also imported into the H5Zblosc module so users can
refer to them as one of:

- H5Zblosc.NOSHUFFLE
- H5Zblosc.SHUFFLE
- H5Zblosc.BITSHUFFLE