Avoid calling non-public h5fd_xxx_init() functions (#967)
* Avoid calling non-public h5fd_xxx_init() functions

The h5fd_xxx_init() functions are not part of the HDF5 public API.  A
method of retrieving the hid_t value of drivers that uses
public/supported API calls was suggested in a comment on a related
github issue and is implemented in this commit.  For more details see:

https://github.com/HDFGroup/hdf5/issues/1809#issuecomment-1154397046

* Initialize fapl minimally in Drivers.__init__()

* Disable file locking in Drivers.__init()

Unless the user has already set `HDF5_USE_FILE_LOCKING` in the
environment, set it to `FALSE` to avoid causing problems with mmap'ing.
This needs to be done before the first call to the HDF5 APIs.

* Change how parallel HDF5 suppport is detected

Instead of detecting parallel HDF5 support by whether `h5fd_mpio_init()`
throws an exception, parallel support is detected by finding the
`H5Pset_fapl_mpio` symbol in any loaded library whose name matches
`r"hdf5"i`.  The documentation for `H5Pset_fapl_mpio()` states:

> H5Pset_fapl_mpio() is available only in the parallel HDF5 library

The detection has also been moved into `Drivers.__init__()`
and the inclusion of `mpio.jl` (upon loading of the MPI package) will
only happen if the HDF5 libraries have parallel support.

* Add DRIVERS entry for MPIO lazily in set_driver!

* Replace `contains()` with call to older `match()`

This maintains compatibility with pre-1.5 versions of Julia.

* Improve usage of Libdl

* Use HDF5.API.libhdf5 rather than dllist()