Fix H5Oexist test result after HDF5 bugfix (#1170)
A bugfix in the HDF5 library changed the return value of the
H5Oexists(_by_name) API call. In earlier versions of the library,
when the ID was a file ID and the object doesn't exist, the call
would return FAIL (-1). The correct behavior is to return false (0).
This change will appear in the next released version of HDF5,
which is planned to be 2.0.0.

This change updates the object test to expect FAIL for 1.14.5 and
earlier and false for later versions.